+++ src/main/java/com/oms/dto/OrderDTO.java
@@ -3,6 +3,7 @@ package com.oms.dto;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Builder;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Builder
@@ -14,11 +15,11 @@ public record OrderDTO(
     
     @NotNull(message = "Amount is required")
     @Positive(message = "Amount must be positive")
-    Double amount,
+    BigDecimal amount,
     
-    Double discountAmount,
+    BigDecimal discountAmount,
     
-    Double finalAmount,
+    BigDecimal finalAmount,
     
     LocalDateTime orderDate
 ) {
